+++ blueprints/cis/level2/main.tf
@@ -11,12 +11,15 @@ module "core" {
   region_key         = var.region_key
   cis_level          = local.cis_level
 
-  parent_compartment_ocid     = var.parent_compartment_ocid
-  enable_delete               = var.enable_delete
-  enable_tagging              = var.enable_tagging
-  enable_tag_defaults         = var.enable_tag_defaults
-  enable_audit_retention      = var.enable_audit_retention
-  audit_retention_period_days = var.audit_retention_period_days
+  parent_compartment_ocid          = var.parent_compartment_ocid
+  enable_delete                    = var.enable_delete
+  enable_tagging                   = var.enable_tagging
+  enable_tag_defaults              = var.enable_tag_defaults
+  enable_audit_retention           = var.enable_audit_retention
+  audit_retention_period_days      = var.audit_retention_period_days
+  cloud_guard_enabled              = var.cloud_guard_enabled
+  cloud_guard_detector_recipe_ids  = var.cloud_guard_detector_recipe_ids
+  cloud_guard_responder_recipe_ids = var.cloud_guard_responder_recipe_ids
 
   defined_tags = var.defined_tags
   freeform_tags = merge(